Switch from idlewait to sleep in the Emscripten port
The semantics are a bit different (we're called every emulated tick),
so we need to take into account the sleep time.

Also add a hook for knowing when disk writes happen, which we can use as
a proxy for then the Mac is done booting (there are a couple of writes
during boot, and then they stop).

Updates mihaip/infinite-mac#71